// frontend/src/enums/user.enum.ts
export enum USER_STATUSES {
  ACTIVE = "active",
  INACTIVE = "inactive",
  DELETED = "deleted",
}

export enum USER_ROLES {
  ADMIN = "admin",
  SELLER = "seller",
  BIDDER = "bidder",
}

export enum USER_ACTIONS {
  DELETE = "delete",
  RESTORE = "restore",
  DEACTIVATE = "deactivate",
  ACTIVATE = "activate",
}